Add support for single-module projects
This started out simple. :)

First thing I realized was that with the module argument becoming
optional (and so an option, as in `--module $MODULE`), the only
remaining positional argument is the version. So for consistency I moved
that into a required `--version $VERSION` option. **This is a breaking
change.**

I wanted to test this against a project without modules. I looked for a
likely candidate in the recent VOTE threads on the ASF incubator lists.
Ended up choosing Dubbo. Problem is, the filename of the `.zip` of
Dubbo, AND ALSO the directory inside the archive, are not named exactly
like their Zipkin counterparts. After hacking around with flags, I
decided to do this properly and introduce a little templating for this.
As usual, the defaults work for Zipkin.  So, testing against Dubbo goes
like this:

```
env BUILD_IMAGE=1 ./check.sh \
    --project dubbo --version 2.7.1 --gpg-key 9C44B5A6884984DB \
    --git-hash a36cc7520e0150d07d30e8baa0c61d5f9f11e6ed \
    --incubating --repo release \
    --zipname-template 'apache-{project}{dash_module}{dash_incubating}-{version}-src' \
    --sourcedir-template 'apache-{project}{dash_module}{dash_incubating}-{version}-src'
```

Notice the absence of a `--module` argument.

Running against a Zipkin module:

```
env BUILD_IMAGE=1 ./check.sh --module brave-cassandra \
    --version 0.10.2 --gpg-key BB67A050 \
    --git-hash b8e379f33eb9f52f0742d3187252de330163cbf7 \
    --repo release
```

While doing this I noticed that if an intermediate step fails (like
unzipping), then down the line we might crash with an exception (like
when specifying some non-existent directory as the workdir to run a
subprocess). So I also wrapped each check in an exception handler.