Fix #25: Introduce multiple levels of failures
Highlights:
* Can now differentiate between "check failed" and "failed to run check"
(latter gets printed as `[ERROR]`)
* Individual checks can return various error levels, depending on the
severity of problems they find. For example, the build-and-test check
reports a `FAIL` level problem if it knows how to run tests, but they
fail; on the other hand it reports a `NOTE` level problem if it doesn't
know how to run tests.
* Inline problem reporting is correctly colorized by level
* Repo-Release archive comparison is downgraded to `WARN` until #20 is
done